+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+## [3.4.26](https://github.com/KongHack/FormConfig/releases/tag/3.4.26)
+ - @GameCharmer add removeField method to FormConfig
+ - @GameCharmer allow passing a callable for post-processing in Make Read Only
+
+
+
 ## [3.4.25](https://github.com/KongHack/FormConfig/releases/tag/3.4.25)
  - @GameCharmer Add same `|raw` argument to the HTML version of select2